feat: io tests
